updated for UI issues

+++ potential_app/models.py
@@ -89,8 +89,16 @@ class Land(models.Model):
         related_name='lands'
     )
     name = models.CharField(max_length=100)
-    latitude = models.FloatField()
-    longitude = models.FloatField()
+
+    latitude = models.DecimalField(
+        max_digits=9,
+        decimal_places=6
+    )
+    longitude = models.DecimalField(
+        max_digits=9,
+        decimal_places=6
+    )
+
     area_m2 = models.FloatField()
     address = models.TextField(blank=True)
     proof_document = models.FileField(
@@ -99,6 +107,7 @@ class Land(models.Model):
         blank=True
     )
     created_at = models.DateTimeField(auto_now_add=True)
+
 
     def __str__(self):
         return self.name
@@ -123,8 +132,8 @@ class LandAnalysis(models.Model):
         related_name='analyses'
     )
 
-    latitude = models.FloatField()
-    longitude = models.FloatField()
+    latitude = models.DecimalField()
+    longitude = models.DecimalField()
     area_m2 = models.FloatField(null=True, blank=True)
     area_ha = models.FloatField(null=True, blank=True)
